Connexion, validation candidature...

+++ Interface/Sources/ajax.php
@@ -9,6 +9,9 @@
 			$_SESSION['groupId']
 		){
 		protect(ELEVE);
+		$group = getGroupFromGroupId($_SESSION['groupId']);
+		if($group['EtatCandidature']!=1)
+			die("La candidature est déjà validée, il est impossible de modifier l'ordre des choix.");
 		$order = explode(';', $_POST['order']);
 		$db->query('DELETE FROM ChoixGroupe WHERE idG='.$_SESSION['groupId']) or die(mysqli_error($db));;
 		$values = '';
@@ -17,5 +20,10 @@
 			$values .= ($count?',':'').'('.$idProj.', '.$_SESSION['groupId'].', '.(++$count).')';
 		
 		$db->query('INSERT INTO ChoixGroupe (idProj, idG, `index`) VALUES '.$values) or die(mysqli_error($db));;
+	}else if(isset($_POST['action']) && $_POST['action']=='confirm-choices' && $_SESSION['groupId']){
+		$db->query('DELETE FROM ChoixGroupe WHERE `index` > 6 AND idG='.$_SESSION['groupId']) or die(mysqli_error($db));;
+		$db->query('UPDATE Groupe SET EtatCandidature=2 WHERE idG='.$_SESSION['groupId']) or die(mysqli_error($db));;
+	}else{
+		echo 'invalid request';
 	}
 ?>
